Resize potato-ham-soup.jpg
The image was too large to be viewed properly on my computer screen. I
imagine there is a better fix for these things with CSS and that it may
be better to specify the original height and width only, unlike what I
did, stretching the image. This is a band-aid solution.